// src/app/shared/models/nutrition-data.interface.ts
export interface INutritionData {
  calories: number;
  totalFat: number;
  saturatedFat: number;
  transFat: number;
  polyunsaturatedFat: number;
  monounsaturatedFat: number;
  totalCarbohydrate: number;
  sugar: number;
  fiber: number;
  protein: number;
  cholesterol: number;
  sodium: number;
}